refactor: 카테고리 리팩터링 리뷰 반영 #758 (#792)
* style: reCreate -> recreate로 수정

* refactor: FileUiModel -> UploadFile로 이름 변경

* style: 날짜 Long 한의 이해를 돕기 위한 주석 추가

* fix: 카테고리가 비어있는 상태에서 스타카토 생성 버튼을 클릭하면 크래시가 발생하는 오류 수정

* feat: 카테고리 삭제 시 id가 null일 경우 안내 메시지 표시

* refactor: isDeleteSuccess -> isDeleted로 이름 변경

* feat: CategoryUpdateError에 state 프로퍼티 추가

+++ android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/category/viewmodel/CategoryViewModel.kt
@@ -35,8 +35,8 @@ class CategoryViewModel
         private val _exceptionState = MutableSingleLiveData<ExceptionState2>()
         val exceptionState: SingleLiveData<ExceptionState2> get() = _exceptionState
 
-        private val _isDeleteSuccess = MutableSingleLiveData<Boolean>(false)
-        val isDeleteSuccess: SingleLiveData<Boolean> get() = _isDeleteSuccess
+        private val _isDeleted = MutableSingleLiveData<Boolean>()
+        val isDeleted: SingleLiveData<Boolean> get() = _isDeleted
 
         fun loadCategory(id: Long) {
             if (id <= DEFAULT_CATEGORY_ID) {
@@ -54,7 +54,12 @@ class CategoryViewModel
 
         fun deleteCategory() {
             viewModelScope.launch {
-                val id = _category.value?.id ?: return@launch
+                val id = _category.value?.id
+                if (id == null) {
+                    _isDeleted.setValue(false)
+                    return@launch
+                }
+
                 val result: ApiResult<Unit> = categoryRepository.deleteCategory(id)
                 result.onSuccess { updateIsDeleteSuccess() }
                     .onServerError(::handleServerError)
@@ -67,7 +72,7 @@ class CategoryViewModel
         }
 
         private fun updateIsDeleteSuccess() {
-            _isDeleteSuccess.setValue(true)
+            _isDeleted.setValue(true)
         }
 
         private fun handleServerError(message: String) {
